requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv